Refactor AI Assist package structure and enhance UI components
- Organize AI Assist classes into specialized sub-packages: `overlay`, `validation`, `error`, and `diff`
- Update `AiAssistOverlay` to use `getSpriteFromXmlString` for clipboard processing and update related imports
- Enhance `AiTutorErrorDialog` by applying a `RoundedCornerShape` to the `AlertDialog`
- Relocate `AiTutorSpriteValidator`, `AiAssistOverlay`, `AiAssistOverlayCallbacks`, and `AiTutorErrorDialog` to their respective new packages
